requirements.txt
langchain
langchain-openai
langgraph
playwright
python-dotenv
pydantic
aiohttp
aiofiles
beautifulsoup4
